akamai/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
edgegrid-python==1.1.1
enum34==1.1.6
httpie==1.0.2
httpie-edgegrid==1.0.2
idna==2.8
ipaddress==1.0.22
ndg-httpsclient==0.5.1
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.0
pyOpenSSL==19.0.0
PyYAML==5.1
requests==2.21.0
six==1.12.0
urllib3==1.24.2
